Refactored TextbitEditable and added yjs and remote cursor handling
