+++ package.json
@@ -45,6 +45,7 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "react": "^18.2.0",
+    "react-icons": "^5.0.1",
     "rollup": "^4.13.0",
     "rollup-plugin-dts": "^6.1.0",
     "rollup-plugin-import-css": "^3.5.0",
@@ -54,11 +55,12 @@
     "ts-node": "^10.9.2",
     "tslib": "^2.6.2",
     "typescript": "^5.4.2",
-    "react-icons": "^5.0.1"
+    "y-protocols": "^1.0.6"
   },
   "peerDependencies": {
+    "react": "^18.2.0",
     "@slate-yjs/core": "^1.0.2",
-    "react": "^18.2.0"
+    "y-protocols": "^1.0.6"
   },
   "dependencies": {
     "@slate-yjs/react": "^1.1.0",
