+++ src/components/TextbitEditable/TextbitEditable.tsx
@@ -2,22 +2,18 @@ import React, { // Necessary for esbuild
   PropsWithChildren,
   useEffect,
   useMemo,
-  useCallback,
-  useContext
+  useCallback
 } from 'react'
-import { createEditor, Editor as SlateEditor, Descendant, Transforms, Element as SlateElement, Range, Path, Node, BaseEditor, Editor } from "slate"
+import { createEditor, Editor as SlateEditor, Descendant, BaseEditor } from "slate"
 import { HistoryEditor, withHistory } from "slate-history"
-import { Editable, ReactEditor, RenderElementProps, RenderLeafProps, Slate, useFocused, withReact } from "slate-react"
-import * as uuid from 'uuid'
-import { YHistoryEditor } from '@slate-yjs/core'
+import { ReactEditor, RenderElementProps, RenderLeafProps, withReact } from "slate-react"
+import { WithCursorsOptions, YHistoryEditor } from '@slate-yjs/core'
 
 import { DragStateProvider } from './DragStateProvider'
 import { withInline } from './with/inline'
-import { calculateStats } from '@/lib/index'
 
 import { ElementComponent } from './components/Element'
 import { Leaf } from './components/Leaf'
-import { toggleLeaf } from '@/lib/toggleLeaf'
 import { withInsertText } from './with/insertText'
 import { withNormalizeNode } from './with/normalizeNode'
 import { withEditableVoids } from './with/editableVoids'
@@ -25,42 +21,57 @@ import { withInsertBreak } from './with/insertBreak'
 import { withInsertHtml } from './with/insertHtml'
 import { PresenceOverlay } from './components/PresenceOverlay/PresenceOverlay'
 import { useTextbit } from '../TextbitRoot'
-import { debounce } from '@/lib/debounce'
 import { usePluginRegistry } from '../PluginRegistry'
-import { PluginRegistryAction, PluginRegistryComponent } from '../PluginRegistry/lib/types'
 import { PositionProvider } from '../ContextTools/PositionProvider'
 import { Gutter } from '../GutterProvider'
-import { FocusContext } from '../TextbitRoot/FocusContext'
+import { Awareness } from 'y-protocols/awareness'
+import * as Y from 'yjs'
+import { SlateEditable } from './components/Slate/SlateEditable'
+import { SlateSlate } from './components/Slate/SlateSlate'
+import { useYjsEditor } from '@/hooks'
 
-
-export const TextbitEditable = ({ children, value, onChange, yjsEditor, gutter = true, dir = 'ltr', className = '' }: PropsWithChildren & {
+interface TextbitEditableProps extends PropsWithChildren {
   onChange?: (value: Descendant[]) => void
   value?: Descendant[]
-  yjsEditor?: SlateEditor
-  gutter?: boolean
+  gutter?: boolean // FIXME: This might not be used no longer
   dir?: 'ltr' | 'rtl'
   className?: string
-}) => {
-  const inValue = value || [{
-    id: uuid.v4(),
-    type: "core/text",
-    class: "text",
-    children: [
-      { text: "" }
-    ]
-  }]
+  sharedRoot?: Y.XmlText
+  awareness?: Awareness
+  cursorOptions?: WithCursorsOptions
+}
 
-  const {
-    dispatch,
-    debounce: debounceTimeout,
-    placeholders: displayPlaceholders
-  } = useTextbit()
+interface TextbitEditableBaseProps extends PropsWithChildren {
+  onChange: (value: Descendant[]) => void
+  value: Descendant[]
+}
 
-  const {
-    plugins,
-    components,
-    actions
-  } = usePluginRegistry()
+interface TextbitEditableYjsProps extends PropsWithChildren {
+  sharedRoot: Y.XmlText
+}
+
+interface TextbitEditableYjsAwarenessProps extends PropsWithChildren {
+  sharedRoot: Y.XmlText
+  awareness: Awareness
+  cursorOptions: WithCursorsOptions
+}
+
+export const TextbitEditable = ({
+  children,
+  value,
+  onChange,
+  sharedRoot,
+  awareness,
+  cursorOptions,
+  gutter = true,
+  dir = 'ltr',
+  className = ''
+}:
+  TextbitEditableProps & (TextbitEditableBaseProps | TextbitEditableYjsProps | TextbitEditableYjsAwarenessProps)
+) => {
+  const yjsEditor = useYjsEditor({ sharedRoot, awareness, cursorOptions })
+  const { placeholders } = useTextbit()
+  const { plugins, components, actions } = usePluginRegistry()
 
   const textbitEditor = useMemo<BaseEditor & ReactEditor & HistoryEditor>(() => {
     const e = SlateEditor.isEditor(yjsEditor) ? yjsEditor : createEditor()
@@ -81,14 +92,6 @@ export const TextbitEditable = ({ children, value, onChange, yjsEditor, gutter =
     return e
   }, [])
 
-  useEffect(() => {
-    const [words, characters] = calculateStats(textbitEditor)
-    dispatch({
-      words,
-      characters
-    })
-  }, [])
-
 
   const renderSlateElement = useCallback((props: RenderElementProps) => {
     return ElementComponent(props)
@@ -99,55 +102,14 @@ export const TextbitEditable = ({ children, value, onChange, yjsEditor, gutter =
     return Leaf(props)
   }, [])
 
-  // Non debounce onChange handler
-  const directOnChange = (value: Descendant[]) => {
-    if (onChange) {
-      onChange(value)
-    }
-
-    const [words, characters] = calculateStats(textbitEditor)
-
-    dispatch({
-      words,
-      characters
-    })
-  }
-
-  // Debounce onChange handler
-  const debouncedOnchange = useMemo(() => {
-    return debounce((value: Descendant[]) => {
-      directOnChange(value)
-    }, debounceTimeout)
-  }, [])
-
-  // Handle onchange and use direct or debounced handler
-  const handleOnChange = useCallback((value: Descendant[]) => {
-    const isAstChange = textbitEditor.operations.some(
-      op => 'set_selection' !== op.type
-    )
-
-    if (isAstChange) {
-      if (debounceTimeout !== 0) {
-        debouncedOnchange(value)
-      }
-      else {
-        directOnChange(value)
-      }
-    }
-  }, [])
-
   return (
     <DragStateProvider>
-      <Slate
-        editor={textbitEditor}
-        initialValue={inValue}
-        onChange={(value) => { handleOnChange(value) }}
-      >
+      <SlateSlate editor={textbitEditor} value={value} onChange={onChange}>
         <PositionProvider inline={true}>
           <Gutter.Provider dir={dir} gutter={gutter}>
 
             <Gutter.Content>
-              <PresenceOverlay isCollaborative={!!yjsEditor}>
+              <PresenceOverlay isCollaborative={!!awareness}>
                 <SlateEditable
                   className={className}
                   renderSlateElement={renderSlateElement}
@@ -155,7 +117,7 @@ export const TextbitEditable = ({ children, value, onChange, yjsEditor, gutter =
                   textbitEditor={textbitEditor}
                   actions={actions}
                   components={components}
-                  displayPlaceholders={displayPlaceholders}
+                  displayPlaceholders={placeholders}
                 />
               </PresenceOverlay>
             </Gutter.Content>
@@ -164,97 +126,7 @@ export const TextbitEditable = ({ children, value, onChange, yjsEditor, gutter =
 
           </Gutter.Provider>
         </PositionProvider>
-      </Slate>
+      </SlateSlate>
     </DragStateProvider >
   )
-}
-
-const SlateEditable = ({ className, renderSlateElement, renderLeafComponent, textbitEditor, actions, components, displayPlaceholders }: {
-  className: string
-  renderSlateElement: (props: RenderElementProps) => JSX.Element
-  renderLeafComponent: (props: RenderLeafProps) => JSX.Element
-  textbitEditor: Editor
-  actions: PluginRegistryAction[]
-  components: Map<string, PluginRegistryComponent>
-  displayPlaceholders: boolean
-}): JSX.Element => {
-  const focused = useFocused()
-  const { setFocused } = useContext(FocusContext)
-
-  useEffect(() => {
-    setFocused(focused)
-  }, [focused])
-
-  return (
-    <Editable
-      data-state={focused ? 'focused' : ''}
-      className={className}
-      renderElement={renderSlateElement}
-      renderLeaf={renderLeafComponent}
-      onKeyDown={event => handleOnKeyDown(textbitEditor, actions, event)}
-      decorate={([node, path]) => handleDecoration(textbitEditor, components, node, path, displayPlaceholders)}
-    />
-  )
-}
-
-/*
- * Display decoration when node is
- * 1. not the editor
- * 2. node is empty
- * 3. selection is on this node
- * 4. selection is collapsed (it does not span more nodes)
- */
-function handleDecoration(editor: SlateEditor, components: Map<string, PluginRegistryComponent>, node: Node, path: Path, displayPlaceholders: boolean) {
-  if (
-    editor.selection != null &&
-    !SlateEditor.isEditor(node) &&
-    SlateEditor.string(editor, [path[0]]) === "" &&
-    Range.includes(editor.selection, path) &&
-    Range.isCollapsed(editor.selection) &&
-    SlateElement.isElement(node)
-  ) {
-    const entry = components.get(node.type)
-
-    return [
-      {
-        ...editor.selection,
-        placeholder: (entry?.componentEntry?.placeholder && displayPlaceholders) ? entry.componentEntry.placeholder : ''
-      }
-    ]
-  }
-
-  return []
-}
-
-/*
- * Match key events to registered actions keyboard shortcuts. Then either
- * 1. call their action handler
- * 2. toggle leafs on or off
- * 3. transform text nodes to another type
- */
-function handleOnKeyDown(editor: SlateEditor, actions: PluginRegistryAction[], event: React.KeyboardEvent<HTMLDivElement>) {
-  for (const action of actions) {
-    if (!action.isHotkey(event)) {
-      continue
-    }
-
-    event.preventDefault()
-
-    if (action.handler && true !== action.handler({ editor })) {
-      break
-    }
-
-    if (action.plugin.class === 'leaf') {
-      toggleLeaf(editor, action.plugin.name)
-    }
-    else if (action.plugin.class === 'text') {
-      // FIXME: Should not allow transforming blocks (only text class element)
-      Transforms.setNodes(
-        editor,
-        { type: action.plugin.name },
-        { match: n => SlateElement.isElement(n) && SlateEditor.isBlock(editor, n) }
-      )
-    }
-    break
-  }
 }
