+++ src/hooks/index.tsx
@@ -1,3 +1,4 @@
 export { useClickGlobal } from './useClickGlobal'
 export { useKeydownGlobal } from './useKeyGlobal'
 export { useMouseMove } from './useMouseMove'
+export { useYjsEditor } from './useYjsEditor'
